made the code more readable, added utils, created execution.c file
