+++ out
@@ -1,0 +1,114 @@
+Libft
+Makefile
+README.md
+cd.c
+checks.c
+echo.c
+env.c
+execution.c
+f1
+f2
+freeing.c
+ft_printf
+history.c
+inpt_functions.c
+loop.c
+minishell
+minishell.c
+minishell.dSYM
+minishell.h
+out
+outfile
+parsing.c
+pipe.c
+pipex_functions.c
+pwd.c
+set_nodes.c
+sig_functions.c
+tokens.c
+utils.c
+white_space.c
+total 808
+drwxr-xr-x  83 tsimitop  2022_heilbronn   2822 May  7 21:12 .
+drwxr-xr-x  35 tsimitop  2022_heilbronn   1190 May  7 21:19 ..
+-rw-r--r--   1 tsimitop  2022_heilbronn   1722 Apr 26 17:10 Makefile
+-rwxr-xr-x   1 tsimitop  2022_heilbronn  51896 May  2 15:28 a.out
+-rw-r--r--   1 tsimitop  2022_heilbronn   1554 Apr 25 17:33 ft_atoi.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    956 May  7 21:12 ft_atoi.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1292 Apr 25 17:33 ft_bzero.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    676 May  7 21:12 ft_bzero.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1704 Apr 25 17:33 ft_calloc.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    768 May  7 21:12 ft_calloc.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1159 Apr 25 17:33 ft_isalnum.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    720 May  7 21:12 ft_isalnum.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1382 Apr 25 17:33 ft_isalpha.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    680 May  7 21:12 ft_isalpha.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1288 Apr 25 17:33 ft_isascii.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    712 May  7 21:12 ft_isascii.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1120 Apr 25 17:33 ft_isdigit.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    664 May  7 21:12 ft_isdigit.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1274 Apr 25 17:33 ft_isprint.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    712 May  7 21:12 ft_isprint.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   2856 Apr 25 17:33 ft_itoa.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   2280 May  7 21:12 ft_itoa.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1145 Apr 25 17:33 ft_lstadd_back.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    996 Apr 25 17:33 ft_lstadd_front.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1091 Apr 25 17:33 ft_lstclear.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1016 Apr 25 17:33 ft_lstdelone.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1026 Apr 25 17:33 ft_lstiter.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1076 Apr 25 17:33 ft_lstlast.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1364 Apr 25 17:33 ft_lstmap.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1330 Apr 25 17:33 ft_lstnew.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1114 Apr 25 17:33 ft_lstsize.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1631 Apr 25 17:33 ft_memchr.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    716 May  7 21:12 ft_memchr.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1708 Apr 25 17:33 ft_memcmp.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    748 May  7 21:12 ft_memcmp.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   2107 Apr 25 17:33 ft_memcpy.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    756 May  7 21:12 ft_memcpy.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1798 Apr 25 17:33 ft_memmove.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    920 May  7 21:12 ft_memmove.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1332 Apr 25 17:33 ft_memset.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    684 May  7 21:12 ft_memset.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1063 Apr 25 17:33 ft_putchar_fd.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    680 May  7 21:12 ft_putchar_fd.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1103 Apr 25 17:33 ft_putendl_fd.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    832 May  7 21:12 ft_putendl_fd.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1349 Apr 25 17:33 ft_putnbr_fd.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1048 May  7 21:12 ft_putnbr_fd.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1080 Apr 25 17:33 ft_putstr_fd.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    720 May  7 21:12 ft_putstr_fd.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   2090 May  2 15:28 ft_split.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1924 May  7 21:12 ft_split.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1536 Apr 25 17:33 ft_strchr.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    748 May  7 21:12 ft_strchr.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1659 Apr 25 17:33 ft_strdup.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    880 May  7 21:12 ft_strdup.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1277 Apr 25 17:33 ft_striteri.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    696 May  7 21:12 ft_striteri.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1462 May  3 17:24 ft_strjoin.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    992 May  7 21:12 ft_strjoin.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   2234 Apr 25 17:33 ft_strlcat.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    984 May  7 21:12 ft_strlcat.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   2266 Apr 25 17:33 ft_strlcpy.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1080 May  7 21:12 ft_strlcpy.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1175 Apr 25 17:33 ft_strlen.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    676 May  7 21:12 ft_strlen.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1346 Apr 25 17:33 ft_strmapi.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    856 May  7 21:12 ft_strmapi.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1773 Apr 25 17:33 ft_strncmp.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    808 May  7 21:12 ft_strncmp.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   2331 Apr 25 17:33 ft_strnstr.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    912 May  7 21:12 ft_strnstr.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1520 Apr 25 17:33 ft_strrchr.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    760 May  7 21:12 ft_strrchr.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   2171 May  2 15:33 ft_strtrim.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1688 May  7 21:12 ft_strtrim.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1507 May  6 12:51 ft_substr.c
+-rw-r--r--   1 tsimitop  2022_heilbronn   1092 May  7 21:12 ft_substr.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1334 Apr 25 17:33 ft_tolower.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    656 May  7 21:12 ft_tolower.o
+-rw-r--r--   1 tsimitop  2022_heilbronn   1010 Apr 25 17:33 ft_toupper.c
+-rw-r--r--   1 tsimitop  2022_heilbronn    656 May  7 21:12 ft_toupper.o
+-rw-r--r--   1 tsimitop  2022_heilbronn  34248 May  7 21:12 libft.a
+-rw-r--r--   1 tsimitop  2022_heilbronn   3129 May  2 15:28 libft.h
