+++ utils.c
@@ -6,7 +6,7 @@
 /*   By: tsimitop <tsimitop@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/25 16:08:36 by tsimitop          #+#    #+#             */
-/*   Updated: 2024/05/07 15:06:21 by tsimitop         ###   ########.fr       */
+/*   Updated: 2024/05/07 21:15:29 by tsimitop         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,4 +43,65 @@ int	skip_whitespace(char *inpt, int i)
 	 inpt[i] == '\n' || inpt[i] == '\v' || inpt[i] == '\f'))
 	 	i++;
 	return (i);
+}
+
+bool	is_metacharacter(char c)
+{
+	if (c == '|' || c =='>' || c == '<' || c == '"' || c == '\'')
+		return (true);
+	return (false);
+}
+
+bool	is_ws(char c)
+{
+	if (c == ' ' || c == '\t')
+		return (true);
+	return (false);
+}
+
+t_command	*get_last_cmd(t_command *cmd)
+{
+	if (!cmd)
+		return (NULL);
+	while (cmd && cmd->next)
+		cmd = cmd->next;
+	return (cmd);
+}
+
+int	num_of_remaining_cmds(t_command *cur) //check if needed
+{
+	int	counter;
+
+	counter = 0;
+	while (cur)
+	{
+		cur = cur->next;
+		counter++;
+	}
+	return (counter);
+}
+
+int	num_of_total_cmds(t_command *cur)
+{
+	int	counter;
+
+	counter = 0;
+	while (cur)
+	{
+		cur = cur->next;
+		counter++;
+	}
+	return (counter);
+}
+
+void	close_fds(t_command *cur)
+{
+	if (cur->input_fd != -1)
+		close(cur->input_fd);
+	if (cur->output_fd != -1)
+		close(cur->output_fd);
+	if (cur->fd[0] != -1)
+		close(cur->fd[0]);
+	if (cur->fd[1] != -1)
+		close(cur->fd[1]);
 }
